Reorganizacion, imports y comentarios de Categorias
