+++ src/cazaVinchucas/Categoria/Experto.java
@@ -1,4 +1,12 @@
-package cazaVinchucas;
+package cazaVinchucas.Categoria;
+
+/**
+ * Clase encargada de representar el comportamiento particular de una 
+ * categoria de usuario experto.
+ * 
+ * @author ivanapr
+ *
+ */
 
 public class Experto extends Categoria {	
 	/**
